Adding group support
In addition to grouping by provider (Vagrant) and provider (VMware), we also try to decode the hostname, and if its WSR-1, make groups for the project, environment, instance (if used) and node purpose as well.

This is now essentially feature complete - but needs significant refactoring to split into functions etc. A config file should be added, PSR-2 should be applied and ideally tests added.