Sorting out warnings etc.
It isn't possible to update the warnings array without returning from the relevant function. This would mean returning a more complex data structure (e.g. an array with valid and invalid messages for get_valid_messages()) and make the calling sites look more complicated.

Ideally this would be replaced with exceptions etc. in the future, but I would rather do this until the choice of language is settled (i.e. keep using PHP or adapt to Python).

Instead we now use a debug flag to output anything that would have been a warning. These debug messages are outputted as they arise.

Importantly these debug messages are not escaped as comments, so running debug mode will not produce a valid Ansible inventory. However this is not an intended use case - its expected where debug mode is used that the user will be reviewing the output to track down bugs etc.